Move property path code to static helper
